requirements.txt
langchain
langgraph
groq
tavily-python
streamlit
pydantic
python-dotenv
reportlab
fpdf2